pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nflreadpy>=0.1.3",
    "openai-agents[redis,voice]>=0.3.3",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
]
